Update BigNumber.java
added modulus and div methods. they continue to execute even though the b<a condition should fail. 
for example, in the modulus method, if a = 100 and b = 30, this method will continue executing even after a = 10 (which is 100- 3*b). However, if you start off with a = 10, b = 30, the method will correctly return 10 as the remainder.